@@ -2,7 +2,7 @@ import Contact from './components/Contact';
 import Nav from './components/Nav';
 import Portfolio from './components/Portfolio';
 import './styles/style.css'
-const { default: Main } = require("./components/Main");
+import Main from "./components/Main";
 
 function App() {
   return (
@@ -15,5 +15,4 @@ function App() {
   );
 }
 
-
-export default App;
+export default App
